+++ packages/core/src/format/interpolate.ts
@@ -1,45 +1,15 @@
-import type { InterpolateOptions, Placeholder, Text, Value } from '../main'
-import { formatDate, formatNumber } from './formatValues'
+import { localizeValue, type InterpolateOptions, type Placeholder, type Text, type Value, ValueWithOptions } from '../main'
 
-function interpolate(
+function interpolate<T>(
   text: Text,
   placeholder: Placeholder,
-  value: Value,
+  value: Value | ValueWithOptions<T>,
   options: InterpolateOptions,
 ): Text {
-  let formattedValue: Text
+  const locale = options.valueLocale
+  const localizedValue = localizeValue(value, locale)
 
-  const isDataObject = typeof value === 'object' && 'date' in value
-  const isNumberObject = typeof value === 'object' && 'number' in value
-
-  if (value instanceof Date) {
-    formattedValue = formatDate(value, options?.datetimeFormat)
-  }
-  else if (isDataObject) {
-    const formatOptions: any = { ...value }
-    delete formatOptions.date
-
-    formattedValue = formatDate(
-      value.date,
-      options?.datetimeFormat,
-      formatOptions,
-    )
-  }
-  else if (isNumberObject) {
-    const formatOptions: any = { ...value }
-    delete formatOptions.number
-
-    formattedValue = formatNumber(
-      value.number,
-      options?.numberFormat,
-      formatOptions,
-    )
-  }
-  else {
-    formattedValue = String(value)
-  }
-
-  return text.replaceAll(`{${placeholder}}`, formattedValue)
+  return text.replaceAll(`{${placeholder}}`, localizedValue)
 }
 
 export default interpolate