+++ packages/core/src/main.ts
@@ -4,22 +4,21 @@ export { default as detectLocaleFromNavigator } from './detection/detectLocaleFr
 export { default as detectLocaleFromPathname } from './detection/detectLocaleFromPathname'
 
 export * from './format/decline'
-export { default as format } from './format/format'
+export * from './format/format'
 export { default as formatToSegments } from './format/formatToSegments'
 export { default as formatToString } from './format/formatToString'
-export * from './format/formatValues'
-export { default as interpolate } from './format/interpolate'
+export * from './format/interpolate'
 
-export * from './localization/format'
 export { default as localizeKey } from './localization/localizeKey'
 export { default as localizeUrl } from './localization/localizeUrl'
+export { default as localizeValue } from './localization/localizeValue'
 
 export { default as collect } from './translation/collect'
 export * from './translation/context'
 export * from './translation/utils'
 
 export * from './constants'
-export * from './utils'
 export { default as initPsitta } from './initPsitta'
+export * from './utils'
 
 export * from './typings'