+++ packages/core/src/utils.ts
@@ -1,20 +1,22 @@
 import { DEFAULT_I18N_CONFIG } from './constants'
-import type { Config, LocaleObject } from './main'
+import type { Config, Locale, LocaleObject } from './main'
+
+export class RelativeTime {
+  constructor(public value: number, public unit: Intl.RelativeTimeFormatUnit) { }
+}
 
 export function getConfig(): Config {
-  const i18nConfig = globalThis.__psitta || DEFAULT_I18N_CONFIG
+  const config: Config = (globalThis.__psitta || DEFAULT_I18N_CONFIG)
   const defaultLocale = getDefaultLocale()
-  const locales = i18nConfig.locales || DEFAULT_I18N_CONFIG.locales
+  const locales = config.locales || DEFAULT_I18N_CONFIG.locales
 
-  const translations = i18nConfig.translations || {}
-  const fallbackLocale = i18nConfig.fallbackLocale || DEFAULT_I18N_CONFIG.fallbackLocale
+  const translations = config.translations || {}
+  const fallbackLocale = config.fallbackLocale || DEFAULT_I18N_CONFIG.fallbackLocale
 
-  const datetimeFormats = i18nConfig.datetimeFormats || DEFAULT_I18N_CONFIG.datetimeFormats
-  const defaultDatetimeFormat = i18nConfig.defaultDatetimeFormat || DEFAULT_I18N_CONFIG.defaultDatetimeFormat
-  const numberFormats = i18nConfig.numberFormats || DEFAULT_I18N_CONFIG.numberFormats
-  const defaultNumberFormat = i18nConfig.defaultNumberFormat || DEFAULT_I18N_CONFIG.defaultNumberFormat
-  const defaultNumberDeclensionRule = i18nConfig.defaultNumberDeclensionRule || DEFAULT_I18N_CONFIG.defaultNumberDeclensionRule
-  const numberDeclensionRules = i18nConfig.numberDeclensionRules
+  const valueLocales = config.valueLocales || DEFAULT_I18N_CONFIG.valueLocales
+  const defaultValueLocale = config.defaultValueLocale || DEFAULT_I18N_CONFIG.defaultValueLocale
+  const defaultNumberDeclensionRule = config.defaultNumberDeclensionRule || DEFAULT_I18N_CONFIG.defaultNumberDeclensionRule
+  const numberDeclensionRules = config.numberDeclensionRules
 
   return {
     translations,
@@ -23,15 +25,13 @@ export function getConfig(): Config {
     defaultLocale,
     defaultNumberDeclensionRule,
     numberDeclensionRules,
-    datetimeFormats,
-    defaultDatetimeFormat,
-    numberFormats,
-    defaultNumberFormat,
+    valueLocales,
+    defaultValueLocale,
   }
 }
 
 export function getDefaultLocale() {
-  const i18nConfig = globalThis.__psitta || DEFAULT_I18N_CONFIG
+  const i18nConfig: Config = globalThis.__psitta || DEFAULT_I18N_CONFIG
   return i18nConfig.defaultLocale || DEFAULT_I18N_CONFIG.defaultLocale
 }
 
@@ -51,4 +51,25 @@ export function parseLocale(lang: string): LocaleObject {
     lang: parsed[0],
     region: parsed[1],
   }
+}
+
+export function getNumberDeclensionRule(locale: Locale, options?: Config) {
+  const { numberDeclensionRules, defaultNumberDeclensionRule }
+    = options || getConfig()
+
+  return numberDeclensionRules?.[locale] || defaultNumberDeclensionRule
+}
+
+export function getValueLocale(locale: Locale, options?: Config) {
+  const { valueLocales, defaultValueLocale } = options || getConfig()
+
+  return valueLocales?.[locale] || defaultValueLocale
+}
+
+export function getFormatOptions(locale: Locale, options?: Config) {
+  const numberDeclensionRule = getNumberDeclensionRule(locale, options)
+  const valueLocale = getValueLocale(locale, options)
+  const formatOptions = { numberDeclensionRule, valueLocale }
+
+  return formatOptions
 }